Create User controller and show view
Added an association between the User and the Event models. A user can
create many events and an event belongs to a creator.

The routes file has the proper routing for the User show page and the
Events index page.

The new users controller contains the show action. It gets the user
based on its id.

The show view contains a heading that indicates the events created by
each user. Then it lists all the events' titles created by said user.
It uses the association mentioned before to access those events.