Add events actions and form
Added logic to new and create actions. New just creates an empty event.
Create makes a new event and receives params from the new private helper
method params_helper. After, it uses the build_ method to create the
association between an event and the creator (a user). Maybe this line
will be changed later.

The form just contains labels for the fields required and the proper
tags to enter the proper values. The 'event_date' field uses a
datetime_local_field to let the user select the day and hour of the
event.

Also added in the events controller a helper method from devise to ask
the user to sign before creating an event.