Add logic to create events
In the events controller, the create action first tries to create an
event object with some parameters passed from the params_required
method. Then, it makes the creator association using the build_creator
method added through the associations between event and user models.
It assigns the current_user.id to prepopulate the proper id. If it's
correctly saved, it redirects to the index page, otherwise it sends a
422 unprocessable entity response.

Both the user and event models contains the has_many and belongs_to
associations to allow the events creation. In the associations, it
specifies the class name and the foreign key because it makes a creator
association instead of the typical user has many events.

The events table needed a foreign key called creator_id, to do that,
there is now a new migration that creates this column and indicates that
it's a foreign key for the user class. Also made the migration to apply
the changes.

And the index view now shows the events names (fixed a typo).