Add logic to let the users attend an event
There is now a new user_events model. This model belongs to both the
user and events models. Also both the user and events model has many
user_events. Finally, to end the associations part, the user model has
many attendees and the event model has many attended_events. This is
done telling rails the new foreign keys and using the keywords source
and through in some of these associations.

Also made the migrations to create the table with the columns and added
the foreign keys needed.

The routes file now contains routes to the new and create actions from
user_events.

There is a new user_events controller to add logic to the new and create
actions. The new action creates an empty user_event record and stores in
a variable all the events. In the create action, it creates another
empty record of user_event, then it builds the attendee association with
the current_user id and the attended_event association with the event id
passed from the "new" view.

The "new" view contains a form with a series of radio buttons. One for
each event, and the user must choose one if it wants to attend and
event. When it clicks the button, it sends this id to the controller to
properly create the user_event record, as explained above.