Add logic to separate attended events
First, in the events controller, there is now a requirement to sign
before seeing and event's page.

The attended events now are separated between those that are classified
as "past events" and those called "future" events. This is done with
the #where and declaring a range inside as argument. Thanks to this
range, we make queries similar to less than or greater than a certain
date. Also store this in a variable. Finally, use a loop to display each
value in an unordereld list, like all the events before this change.