Separate events in index view page
To separate the events like in the show user view, this time we used
first class methods and then scopes. With scopes we allow to return
ActiveRecord instances and keep calling other methods.

In the events controller, there are two variables, one for past events
and other for future events.

Finally in the index view, there are headers to separate the past and
future events and using the variables from the controller to loop and
display each event, similarly like it was before.